Added the first pass at the basic structures
First pass at the basics to get a base that compiles.

This project is going to be all about data structures and algorithms.

Utilizes bitfields heavily.
Individual pieces occupy 7 bits so far.

The Board occupies 64 bytes. Could be shortened to 56;

I have an individual Move down to 4 bytes but this will get smaller. There are more moves stored while computing the next move than any other structure so they need to be as small as possible.

Assuming 1K of working stack room that means that 256 moves can be stored.

On opening, each player has 20 moves available.

Just for guessing: Assuming 20 is the average number of available moves per player, that means that a full evaluation of both sides needs 40 moves worth of room. That means 256 / 40 = 6.4 ply levels ahead might be able to be examined which would be quite good. It would take forever for each move to be determined since we have only one thread, but it would work.

I'm sure I'm overlooking a lot