Major Optimization, details in description
I analyzed the call chain of the four functions involved in the recursive choose_next_move(...) call chain:

* choose_next_move(...)
* the piece move generators
* consider_move(...)
* make_move(...)

and enhanced the piece_gen_t data type to include all of the data that was being allocated as a local variable and/or conputed more than one time in the call chain. Variables such as type, side, whites_turn, col, row and piece are all now just generated once and re-used throughout the call chain!
Also included in the changes to that data type is the change to carry around the game.pieces[] index of the piece being evaluated so that find_piece(...) never has to be called anymore to find the index of the piece being moved!

This speeds up everything that was being computed more than once, as well as adding savings to all of the local variables that are no longer being allocated for each of the 4 functions in the recursive call chain. I am about to re-enable the stats for memory usage and see how much has been trimmed out of the call chain. It may collectively be enough to allow going one ply deeper to ply level 5, with only 2K of flippin' memory lol.