Finally made use of visitor pattern on King's Check state!
Added new king_checker() function that makes use of the visitor pattern I intentionally placed in the piece_gen_t structure on who to call for each move being evaluated.

Now the same mechanism is used to call a local lambda function in the check_kings() function which checks to see if the move places a king in check.